all monitor work moved
